Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main() {
     int r;
-    double pie=3.14;
+    float pie=3.14;
     double area;
     scanf("%f",r);
     area=pie*r*r;
